Refactoring: handle only standard generic MBR
I don't have the tool to handle the other types so I won't be able to
test even if I implement them.

It won't hurt to handle only this type since it's compatible with all
the other types of MBR, it cannot break if you open a different type of
MBR with this library.