Add InvalidSignature error variant; drop Signature55AA type
Allow the caller to cleanly distinguish between a corrupt or missing MBR.
There's no clean way to return an error variant instead of a string when
signature checking happens inside deserialize code, so represent the
signature field as a [u8; 2] instead and do the signature checking in
check().

This is a breaking change.