Zero-fill unused EBR partition entries when writing
The third and fourth EBR entries are defined to contain zeroes, but when
writing we were seeking over them instead of explicitly clearing them.
Since we don't know what was previously stored there (and it might not
have been an EBR) this could result in an EBR containing garbage data.