Add `InvalidSignature` and `InvalidBootFlag` error variants (#18)
* Convert MBRPartitionEntry boot flag to a u8

If we map the u8 to a bool inside the deserializer, there's not a good way
to report a distinct error variant for bad values; we can only report a
string.  Begin refactoring to avoid this.

This is a breaking change.

* Add MBRPartitionEntry.is_active() convenience method

Return true if the boot flag is set.

* Check for invalid boot flag when reading/writing

Callers that want to avoid this can serialize/deserialize the structs
directly with bincode.

* Add InvalidSignature error variant; drop Signature55AA type

Allow the caller to cleanly distinguish between a corrupt or missing MBR.
There's no clean way to return an error variant instead of a string when
signature checking happens inside deserialize code, so represent the
signature field as a [u8; 2] instead and do the signature checking in
check().

This is a breaking change.

* Replace BootstrapCode440 and BootstrapCode446 with serde-big-array

It appears those structs are only needed because serde can't serialize
arrays larger than 32 entries without help.  Rather than carrying the
wrapper code ourselves and exposing it in our API, use regular arrays with
serde-big-array's serialization/deserialization helpers.

Add some additional Eq derives, since we can now.

This is a breaking change.