workflows: update actions; remove monthly scheduled run (#28)
* workflows: remove monthly scheduled run

GitHub disables scheduled workflows on repos that are inactive for two
months, which is very likely to happen here.  As a result, this workflow
would only run twice and then be disabled, taking all of our CI with it.

* workflows: update to current versions of actions

The older versions are using deprecated APIs and throwing warnings.
All of the actions-rs actions are unmaintained, so switch to manually
invoking commands.